Refactor TempUserManager to ensure_agent_role() function
Replace the class-based context manager with a simpler function:
- Rename temp_user_manager.py -> agent_credentials.py
- Replace TempUserManager class with ensure_agent_role() function
- Rename TempUserCredentials -> AgentCredentials
- Update all imports and usages across the codebase
- Update documentation references

The context manager pattern was no longer needed since roles are
persistent (no cleanup) and passwords are deterministic.
